Improve AutoCrafter-Ejection to use items not used for crafting
I wrote some fallback code to eject an item used for crafting (use the ingredient and then eject) but it looks like it is not playing nice with the cachedInventory, so it is commented out.
Looks like the old ejection did have the same problems as mine but I don't use Streams in mine witch may have quit an over-head in this case.