Fixes FabledSkyBlock Stackable Blocks getting reset and not incremented
If pointing a hopper to a stackable, the stackable gets reset to 0 and the item remains in the
hopper.

This is because a `Stackable` instance can be created without any maxSize being set causing it
to default to `0`.
This is probably more of a workaround/hack and should be fixed in FabledSkyBlock...
But it works with right clicking the block manually - So maybe 0 means infinite? I'm not really sure...